adding days of week
