+++ src/js/components/Calendar/stories/Dual.js
@@ -1,58 +1,10 @@
 import React, { Component } from 'react';
 import { storiesOf } from '@storybook/react';
 
-import { Box, Button, Calendar, Grommet, Heading, Text } from 'grommet';
+import { Box, Button, Calendar, Grommet, Heading } from 'grommet';
 import { grommet } from 'grommet/themes';
 
-import {
-  Blank,
-  FormPreviousLink,
-  FormNextLink,
-  Previous,
-  Next,
-} from 'grommet-icons';
-
-class SimpleCalendar extends Component {
-  state = {};
-
-  onSelect = nextDate => {
-    const { date } = this.state;
-    this.setState({ date: nextDate !== date ? nextDate : undefined });
-  };
-
-  render() {
-    const { date } = this.state;
-    return (
-      <Grommet theme={grommet}>
-        <Box align="center" pad="large">
-          <Calendar
-            date={date}
-            onSelect={this.onSelect}
-            size="small"
-            bounds={['2018-09-08', '2018-12-13']}
-          />
-        </Box>
-        <Box align="center" pad="large">
-          <Calendar
-            date={date}
-            daysOfWeek
-            onSelect={this.onSelect}
-            size="small"
-            bounds={['2018-09-08', '2018-12-13']}
-          />
-        </Box>
-      </Grommet>
-    );
-  }
-}
-
-const RangeCalendar = () => (
-  <Grommet theme={grommet}>
-    <Box align="center" pad="large">
-      <Calendar range />
-    </Box>
-  </Grommet>
-);
+import { Blank, Previous, Next } from 'grommet-icons';
 
 const now = new Date();
 const next = new Date(now);
@@ -161,76 +113,4 @@ class DualCalendar extends Component {
   }
 }
 
-class CustomHeaderCalendar extends Component {
-  state = {};
-
-  onSelect = nextDate => {
-    const { date } = this.state;
-    this.setState({ date: nextDate !== date ? nextDate : undefined });
-  };
-
-  render() {
-    const { date } = this.state;
-    return (
-      <Grommet theme={grommet}>
-        <Box align="center" pad="large">
-          <Calendar
-            date={date}
-            onSelect={this.onSelect}
-            size="small"
-            bounds={['2018-09-08', '2018-12-13']}
-            header={({
-              date: currentDate,
-              locale,
-              onPreviousMonth,
-              onNextMonth,
-              previousInBound,
-              nextInBound,
-            }) => (
-              <Box direction="row" align="center" justify="between">
-                <Button disabled={!previousInBound} onClick={onPreviousMonth}>
-                  <Box>
-                    <FormPreviousLink />
-                  </Box>
-                </Button>
-                <Text size="small">
-                  <strong>
-                    {currentDate.toLocaleDateString(locale, {
-                      month: 'long',
-                      year: 'numeric',
-                    })}
-                  </strong>
-                </Text>
-                <Button disabled={!nextInBound} onClick={onNextMonth}>
-                  <Box>
-                    <FormNextLink />
-                  </Box>
-                </Button>
-              </Box>
-            )}
-          />
-        </Box>
-      </Grommet>
-    );
-  }
-}
-
-// DSTCalendar has dates specifically chosen to identify issues with
-// crossing the daylight savings time boundary (from California).
-const DSTCalendar = () => (
-  <Grommet theme={grommet}>
-    <Box align="center" pad="large">
-      <Calendar
-        date="2018-11-04T07:00:00.000Z"
-        bounds={['2013-11-06', '2018-12-06']}
-      />
-    </Box>
-  </Grommet>
-);
-
-storiesOf('Calendar', module)
-  .add('Simple', () => <SimpleCalendar />)
-  .add('Range', () => <RangeCalendar />)
-  .add('Dual', () => <DualCalendar />)
-  .add('Custom Header', () => <CustomHeaderCalendar />)
-  .add('Daylight Savings Time', () => <DSTCalendar />);
+storiesOf('Calendar', module).add('Dual', () => <DualCalendar />);
